requirement.txt
absl-py==0.9.0
alembic==1.4.2
amqp==2.6.0
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.3
awscli==1.18.97
Babel==2.8.0
billiard==3.6.3.0
blinker==1.4
botocore==1.17.20
CacheControl==0.12.6
cachetools==4.1.0
celery==4.4.6
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cognitive-face==1.5.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.7
dlib==19.20.0
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
firebase-admin==3.2.1
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Cors==3.0.8
Flask-DebugToolbar==0.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-Security==3.0.0
flask-shell-ipython==0.4.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
gast==0.3.3
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==19.10.0
h5py==2.10.0
httplib2==0.18.1
idna==2.9
imbalanced-learn==0.7.0
importlib-metadata==1.6.1
imutils==0.5.3
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==4.6.11
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.6.1
matplotlib==3.2.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
opencv-python==4.2.0.34
openface==0.2.1
opt-einsum==3.2.1
pandas==1.0.5
passlib==1.7.2
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==1.0.18
protobuf==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pyopenbr==0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
raven==6.10.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
simplegeneric==0.8.1
six==1.14.0
sklearn==0.0
speaklater==1.3
SQLAlchemy==1.3.18
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.47.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
Werkzeug==0.15.0
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
